Add a simple script to generate test sites
While using random sites on the internet is useful for checking real
world benchmarking improvements, canned data is more useful to ensure
that you've got canned test data to use for isolating improvements to
specific subsections.

The script `generate_site.sh` generates a site with two simple layouts,
a `default` layout, and a `post` layout. Two paragraphs of Lorem Ipsum
are used for each post. The script itself takes two parameters: the
number of posts to generate and the directory to generate the site into.